[WIP] Fix test codes
