+++ src/ts_mruby_internal.hpp
@@ -8,6 +8,7 @@
 
 #include <pthread.h>
 
+#include <fstream>
 #include <iostream>
 #include <map>
 #include <string>
@@ -117,9 +118,35 @@ private:
 };
 
 /*
+ * mruby scripts cache
+ */
+class MrubyScriptsCache {
+public:
+  MOCKABLE_ATTR
+  void store(const std::string &filepath) {
+    std::ifstream ifs(filepath);
+    const std::string code((std::istreambuf_iterator<char>(ifs)),
+                           std::istreambuf_iterator<char>());
+
+    scripts_.insert(make_pair(filepath, code));
+  }
+
+  MOCKABLE_ATTR
+  const std::string &load(const std::string &filepath) {
+    return scripts_[filepath];
+  }
+
+#ifdef MOCKING
+  using mock_type = class MrubyScriptsCacheMock;
+#endif // MOCKING
+
+private:
+  std::map<std::string, std::string> scripts_;
+};
+
+/*
  * Getter for thread-shared mruby script cache
  */
-class MrubyScriptsCache;
 MrubyScriptsCache* getInitializedGlobalScriptCache(const std::string& filepath);
 
 /*
